got rid of various warnings

+++ index/src/InvFPDocList.cpp
@@ -18,7 +18,7 @@ InvFPDocList::InvFPDocList() {
 
 //  This hasn't been tested
 InvFPDocList::InvFPDocList(int id, int len){
-  size = pow(2,DEFAULT);
+  size = (int) pow(2,DEFAULT);
   begin = (int*) malloc(size);
   lastid = begin;
   *lastid = -1;
@@ -48,7 +48,7 @@ InvFPDocList::InvFPDocList(int id, int listlen, int* list, int fr, int* ldocid, 
 
 InvFPDocList::InvFPDocList(MemCache* mc, int id, int len){
   READ_ONLY = false;
-  size = pow(2,DEFAULT);
+  size = (int) pow(2,DEFAULT);
   cache = mc;
   begin = cache->getMem(DEFAULT);
   if (!begin)
@@ -66,7 +66,7 @@ InvFPDocList::InvFPDocList(MemCache* mc, int id, int len){
 
 InvFPDocList::InvFPDocList(MemCache* mc, int id, int len, int docid, int location) {
   READ_ONLY = false;
-  size = pow(2,DEFAULT);
+  size = (int) pow(2,DEFAULT);
   cache = mc;
   begin = cache->getMem(DEFAULT);
   lastid = begin;
@@ -166,7 +166,7 @@ bool InvFPDocList::allocMem() {
   if (READ_ONLY)
     return false;
 
-  size = pow(2,DEFAULT);
+  size = (int) pow(2,DEFAULT);
 
   if (hascache) {
     begin = cache->getMem(DEFAULT);
@@ -379,3 +379,5 @@ int InvFPDocList::logb2(int num) {
     return 23; 
   }
 }
+
+
